refactor(widget-manager): modularize and improve widget handling
- Split handler.py into modular files for UI, state, config, and core logic
- Added __widget__ metadata support with optional runtime fields like hidden
- Enforced main/configure structure and default loading behaviors
- Preserved backward compatibility with INI-based overrides
- Environment Upkeeper marked hidden to avoid UI exposure
