+++ Widgets/widget_manager/__init__.py
@@ -1,0 +1,1 @@
+# __init__.py
