+++ Widgets/widget_manager/default_settings.py
@@ -1,0 +1,369 @@
+default_schema_version = "1.0"
+
+global_widget_defaults = {
+    "WidgetManager": {
+        "show_config_window": "False",
+        "old_menu": "False",
+        "enable_all": "True"
+    },
+    "QuickDock": {
+        "enable_quick_dock": "True",
+        "width": "10",
+        "height": "50",
+        "offset_y": "709",
+        "edge": "left",
+        "unlocked": "False",
+        "buttons_per_row": "8"
+    },
+    "QuickDockColor": {
+        "r": "0.6000000238418579",
+        "g": "0.800000011920929",
+        "b": "1.0",
+        "a": "1.0"
+    },
+    "HeroAI": {
+        "enabled": "False",
+        "category": "Bots",
+        "subcategory": "Multiboxing",
+        "icon": "ICON_BRAIN",
+        "quickdock": "True"
+    },
+    "HeroHelper": {
+        "enabled": "False",
+        "category": "Bots",
+        "subcategory": "Helpers",
+        "icon": "ICON_PEOPLE_CARRY",
+        "quickdock": "True"
+    },
+    "Pet Helper": {
+        "enabled": "False",
+        "category": "Bots",
+        "subcategory": "Helpers",
+        "icon": "ICON_DOG",
+        "quickdock": "True"
+    },
+    "Action Queue Monitor": {
+        "enabled": "False",
+        "category": "Coding",
+        "subcategory": "Info",
+        "icon": "ICON_INFO_CIRCLE",
+        "quickdock": "True"
+    },
+    "Agent Info": {
+        "enabled": "False",
+        "category": "Coding",
+        "subcategory": "Info",
+        "icon": "ICON_BOOK",
+        "quickdock": "True"
+    },
+    "Icon Explorer": {
+        "enabled": "False",
+        "category": "Coding",
+        "subcategory": "ImGui",
+        "icon": "ICON_EXPAND",
+        "quickdock": "True"
+    },
+    "Py4GW_DEMO": {
+        "enabled": "False",
+        "category": "Coding",
+        "subcategory": "Info",
+        "icon": "ICON_INFO",
+        "quickdock": "True"
+    },
+    "Environment Upkeeper": {
+        "enabled": "True",
+        "category": "Coding",
+        "subcategory": "Environment",
+        "icon": "ICON_TREE",
+        "quickdock": "False"
+    },   
+    "Compass +": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Overlays",
+        "icon": "ICON_COMPASS",
+        "quickdock": "True"
+    },
+    "Drop Unyielding Aura": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Triggers",
+        "icon": "ICON_EYE_DROPPER",
+        "quickdock": "True"
+    },
+    "Enter character on load": {
+        "enabled": "True",
+        "category": "Gameplay",
+        "subcategory": "Triggers",
+        "icon": "ICON_PLAY_CIRCLE",
+        "quickdock": "True"
+    },
+    "Instance Timer": {
+        "enabled": "True",
+        "category": "Gameplay",
+        "subcategory": "Overlays",
+        "icon": "ICON_CLOCK",
+        "quickdock": "True"
+    },
+    "Inventory +": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Utilities",
+        "icon": "ICON_SHOPPING_BAG",
+        "quickdock": "True"
+    },
+    "LootManager": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Utilities",
+        "icon": "ICON_SHOPPING_CART",
+        "quickdock": "True"
+    },
+    "Mission Map +": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Overlays",
+        "icon": "ICON_MAP",
+        "quickdock": "True"
+    },
+    "PCons": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Triggers",
+        "icon": "ICON_CHEVRON_UP",
+        "quickdock": "True"
+    },
+    "Switch Character": {
+        "enabled": "True",
+        "category": "Gameplay",
+        "subcategory": "Utilities",
+        "icon": "ICON_PEOPLE_ARROWS",
+        "quickdock": "True"
+    },
+    "Resign on enter Map": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Triggers",
+        "icon": "ICON_FLAG",
+        "quickdock": "True"
+    },
+    "Return to outpost on defeat": {
+        "enabled": "True",
+        "category": "Gameplay",
+        "subcategory": "Triggers",
+        "icon": "ICON_HOME",
+        "quickdock": "True"
+    },
+    "Skip Cinematics": {
+        "enabled": "True",
+        "category": "Gameplay",
+        "subcategory": "Triggers",
+        "icon": "ICON_FAST_FORWARD",
+        "quickdock": "True"
+    },
+    "Titles": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Trigers",
+        "icon": "ICON_ID_BADGE",
+        "quickdock": "True"
+    },
+    "Travel": {
+        "enabled": "True",
+        "category": "Gameplay",
+        "subcategory": "Utilities",
+        "icon": "ICON_MAP_MARKED_ALT",
+        "quickdock": "True"
+    },
+    "Vanquish": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Overlays",
+        "icon": "ICON_FLAG_CHECKERED",
+        "quickdock": "True"
+    }
+}
+
+account_widget_defaults = {
+    "WidgetManager": {
+        "show_config_window": "False",
+        "old_menu": "False",
+        "enable_all": "True"
+    },
+    "QuickDock": {
+        "enable_quick_dock": "True",
+        "width": "10",
+        "height": "50",
+        "offset_y": "709",
+        "edge": "left",
+        "unlocked": "False",
+        "buttons_per_row": "8"
+    },
+    "QuickDockColor": {
+        "r": "0.6000000238418579",
+        "g": "0.800000011920929",
+        "b": "1.0",
+        "a": "1.0"
+    },
+    "HeroAI": {
+        "enabled": "False",
+        "category": "Bots",
+        "subcategory": "Multiboxing",
+        "icon": "ICON_BRAIN",
+        "quickdock": "True"
+    },
+    "HeroHelper": {
+        "enabled": "False",
+        "category": "Bots",
+        "subcategory": "Helpers",
+        "icon": "ICON_PEOPLE_CARRY",
+        "quickdock": "True"
+    },
+    "Pet Helper": {
+        "enabled": "False",
+        "category": "Bots",
+        "subcategory": "Helpers",
+        "icon": "ICON_DOG",
+        "quickdock": "True"
+    },
+    "Action Queue Monitor": {
+        "enabled": "False",
+        "category": "Coding",
+        "subcategory": "Info",
+        "icon": "ICON_INFO_CIRCLE",
+        "quickdock": "True"
+    },
+    "Agent Info": {
+        "enabled": "False",
+        "category": "Coding",
+        "subcategory": "Info",
+        "icon": "ICON_BOOK",
+        "quickdock": "True"
+    },
+    "Icon Explorer": {
+        "enabled": "False",
+        "category": "Coding",
+        "subcategory": "ImGui",
+        "icon": "ICON_EXPAND",
+        "quickdock": "True"
+    },
+    "Py4GW_DEMO": {
+        "enabled": "False",
+        "category": "Coding",
+        "subcategory": "Info",
+        "icon": "ICON_INFO",
+        "quickdock": "True"
+    },
+    "Environment Upkeeper": {
+        "enabled": "True",
+        "category": "Coding",
+        "subcategory": "Environment",
+        "icon": "ICON_TREE",
+        "quickdock": "False"
+    },   
+    "Compass +": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Overlays",
+        "icon": "ICON_COMPASS",
+        "quickdock": "True"
+    },
+    "Drop Unyielding Aura": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Triggers",
+        "icon": "ICON_EYE_DROPPER",
+        "quickdock": "True"
+    },
+    "Enter character on load": {
+        "enabled": "True",
+        "category": "Gameplay",
+        "subcategory": "Triggers",
+        "icon": "ICON_PLAY_CIRCLE",
+        "quickdock": "True"
+    },
+    "Instance Timer": {
+        "enabled": "True",
+        "category": "Gameplay",
+        "subcategory": "Overlays",
+        "icon": "ICON_CLOCK",
+        "quickdock": "True"
+    },
+    "Inventory +": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Utilities",
+        "icon": "ICON_SHOPPING_BAG",
+        "quickdock": "True"
+    },
+    "LootManager": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Utilities",
+        "icon": "ICON_SHOPPING_CART",
+        "quickdock": "True"
+    },
+    "Mission Map +": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Overlays",
+        "icon": "ICON_MAP",
+        "quickdock": "True"
+    },
+    "PCons": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Triggers",
+        "icon": "ICON_CHEVRON_UP",
+        "quickdock": "True"
+    },
+    "Switch Character": {
+        "enabled": "True",
+        "category": "Gameplay",
+        "subcategory": "Utilities",
+        "icon": "ICON_PEOPLE_ARROWS",
+        "quickdock": "True"
+    },
+    "Resign on enter Map": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Triggers",
+        "icon": "ICON_FLAG",
+        "quickdock": "True"
+    },
+    "Return to outpost on defeat": {
+        "enabled": "True",
+        "category": "Gameplay",
+        "subcategory": "Triggers",
+        "icon": "ICON_HOME",
+        "quickdock": "True"
+    },
+    "Skip Cinematics": {
+        "enabled": "True",
+        "category": "Gameplay",
+        "subcategory": "Triggers",
+        "icon": "ICON_FAST_FORWARD",
+        "quickdock": "True"
+    },
+    "Titles": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Trigers",
+        "icon": "ICON_ID_BADGE",
+        "quickdock": "True"
+    },
+    "Travel": {
+        "enabled": "True",
+        "category": "Gameplay",
+        "subcategory": "Utilities",
+        "icon": "ICON_MAP_MARKED_ALT",
+        "quickdock": "True"
+    },
+    "Vanquish": {
+        "enabled": "False",
+        "category": "Gameplay",
+        "subcategory": "Overlays",
+        "icon": "ICON_FLAG_CHECKERED",
+        "quickdock": "True"
+    }
+}
