+++ Widgets/widget_manager/ui_floating_menu.py
@@ -1,0 +1,127 @@
+from Py4GWCoreLib import *
+from . import state
+from .handler import handler
+from .ui_widget_menu import draw_widget_popup_menus
+from .ui_embedded_config import draw_embedded_widget_config
+
+def draw_floating_menu():
+    draw_embedded_widget_config()
+    chara_select = Player.InCharacterSelectScreen()
+    ingame_frame_id = UIManager.GetFrameIDByHash(1144678641)
+    login_frame_id = UIManager.GetChildFrameID(2232987037,[0])
+
+    fullscreen_frame_id = UIManager.GetFrameIDByHash(140452905)
+    left, top, right, bottom = UIManager.GetFrameCoords(fullscreen_frame_id)
+    screen_w = right - left
+    screen_h = bottom - top
+
+    content_frame_id = ingame_frame_id
+    
+    if chara_select:
+        content_frame_id = login_frame_id
+    
+    frame_left, frame_top, frame_right, frame_bottom = UIManager.GetFrameCoords(content_frame_id) 
+
+    if chara_select:
+        button_x = frame_right - 8
+        button_y = frame_top - 5
+        base     = (0.20, 0.25, 0.32, 1.0)
+        hover    = (0.30, 0.35, 0.42, 1.0)
+        active   = (0.15, 0.18, 0.24, 1.0)
+        border   = (0.50, 0.55, 0.60, 1.0)
+    elif UIManager.IsWindowVisible(ingame_frame_id):
+        button_x = 20
+        button_y = 20
+        base   = (0.08, 0.08, 0.08, 1.0)
+        hover  = (0.16, 0.16, 0.16, 1.0)
+        active = (0.05, 0.05, 0.05, 1.0)
+        border = (0.25, 0.25, 0.25, 1.0)
+    else:
+        button_x = frame_right - 12
+        button_y = frame_top - 11
+        base     = (0.40, 0.36, 0.33, 1.0)
+        hover    = (0.48, 0.44, 0.40, 1.0)
+        active   = (0.28, 0.26, 0.23, 1.0)
+        border   = (0.85, 0.82, 0.78, 1.0)
+    
+    window_flags = (PyImGui.WindowFlags.NoTitleBar | PyImGui.WindowFlags.AlwaysAutoResize | 
+                    PyImGui.WindowFlags.NoResize | PyImGui.WindowFlags.NoMove | 
+                    PyImGui.WindowFlags.NoScrollbar | PyImGui.WindowFlags.NoBackground)
+    
+    PyImGui.push_style_color(PyImGui.ImGuiCol.Button,        base)
+    PyImGui.push_style_color(PyImGui.ImGuiCol.ButtonHovered, hover)
+    PyImGui.push_style_color(PyImGui.ImGuiCol.ButtonActive,  active)
+    PyImGui.push_style_color(PyImGui.ImGuiCol.Border,        border)
+    PyImGui.push_style_var(ImGui.ImGuiStyleVar.FrameBorderSize, 1.0)
+    PyImGui.push_style_var(ImGui.ImGuiStyleVar.FrameRounding, 3.0)
+    PyImGui.push_style_color(PyImGui.ImGuiCol.Text, (1.0, 1.0, 1.0, 1.0))
+
+    flip_widgets = ((button_x - 80) > screen_w / 2)
+    if flip_widgets:
+        button_x = frame_left - 90
+        state.left_side = True
+    else:
+        state.left_side = False
+
+    PyImGui.set_next_window_pos(button_x, button_y)
+    PyImGui.set_next_window_size(100, 0)
+    
+    if PyImGui.begin("##floating_button", window_flags):
+        button_x, button_y = PyImGui.get_window_pos()
+        menu_y = button_y + 35 
+        menu_x = button_x + 12 
+
+        if state.popup_open and state.popup_height_known:
+            window_y = button_y
+            window_center_y = window_y + 35 / 2
+            if window_center_y < screen_h / 2:
+                menu_y = button_y + 45  # open downward
+                state.opening_downward = True
+            else:
+                menu_y = button_y - state.popup_height # open upward
+                state.opening_downward = False
+                
+        
+        icon = IconsFontAwesome5.ICON_CIRCLE if state.popup_open else IconsFontAwesome5.ICON_DOT_CIRCLE
+            
+        if state.left_side:
+            button_label = f"Widgets {icon}##WigetUIButton"
+        else:
+            button_label = f"{icon} Widgets##WigetUIButton"
+        
+        if PyImGui.button(button_label, 0, 0):
+                PyImGui.open_popup("FloatingMenu")
+                state.popup_open = True
+        PyImGui.set_next_window_pos(menu_x, menu_y)
+        PyImGui.pop_style_color(5)
+        PyImGui.pop_style_var(2)
+        
+        if PyImGui.begin_popup("FloatingMenu"):
+            state.popup_height = PyImGui.get_window_height()
+            state.popup_height_known = True
+            if PyImGui.button(IconsFontAwesome5.ICON_RETWEET + "##Reload Widgets"):
+                ConsoleLog(state.module_name, "Reloading Widgets...", Py4GW.Console.MessageType.Info)
+                initialized = False
+                handler.discover_widgets()
+                initialized = True
+            ImGui.show_tooltip("Reloads all widgets")
+            PyImGui.same_line(0.0, 10)
+            is_enabled = state.enable_all
+            toggle_label = IconsFontAwesome5.ICON_TOGGLE_ON if state.enable_all else IconsFontAwesome5.ICON_TOGGLE_OFF
+            if is_enabled:
+                PyImGui.push_style_color(PyImGui.ImGuiCol.Button, (0.153, 0.318, 0.929, 1.0))
+                PyImGui.push_style_color(PyImGui.ImGuiCol.ButtonHovered, (0.6, 0.6, 0.9, 1.0))
+                PyImGui.push_style_color(PyImGui.ImGuiCol.ButtonActive, (0.6, 0.6, 0.6, 1.0))
+            if PyImGui.button(toggle_label + "##widget_disable"):
+                state.enable_all = not state.enable_all
+                handler._write_global_setting(state.module_name, "state.enable_all", str(state.enable_all))
+            if is_enabled:
+                PyImGui.pop_style_color(3)
+            ImGui.show_tooltip("Toggle all widgets")
+            PyImGui.separator()
+            draw_widget_popup_menus()
+            PyImGui.end_popup()
+        else:
+            state.popup_open = False
+
+    PyImGui.end()
